actually remove save_checkpoint
(cherry picked from commit 291495abfa60aefccca1eb1b2e9d5fb5aa66eaeb)
